Put the quiz detailed report to separate page.
